fix(profiler): correct category breakdown (exclusive self-time) + CPU dtype
The category breakdown double-counted nested modules (parents summed with
their children -> ~300% total). Switch to exclusive self-time attributed to
each module's nearest TIMED ancestor (robust to untimed nn.ModuleList
containers), so categories sum to ~100% and attention's in-module SDPA stays
in the attention bucket. Also default dtype to bfloat16 (float16 ops are
largely unimplemented on CPU) and fix the OOM hint.

Verified on the real Wan 1.3B DiT: attention 60.1%, FFN 38.5%, other ~1%.

+++ src/wan_profiler/config.py
@@ -68,7 +68,7 @@ class ProfileConfig:
     model_path: Optional[str] = None
     output_dir: str = "results"
     low_memory: bool = True
-    dtype: str = "float16"
+    dtype: str = "bfloat16"  # CPU-safe; float16 ops are largely unimplemented on CPU
     profile_time: bool = True
     profile_flops: bool = False
     profile_memory: bool = False
